requirements.txt
google-genai
python-dotenv
pytest
pyrefly
